requirements.txt
cbpro # The Python client for the Coinbase Pro API (formerly known as the GDAX)
confluent_kafka
ccxt
websocket-client-py3
krakenex
btfxwss